Split JSGen into JSGen and SJSGen
This will be necessary as VarGen will need some low-level things from
JSGen (notably genLet).
